[refactor][ux] improved how captain switching is handled
