+++ app/src/main/java/com/example/werewolfofthemillershollow/turn/CaptainTurn.kt
@@ -6,6 +6,7 @@ import com.example.werewolfofthemillershollow.GameActivity
 import com.example.werewolfofthemillershollow.R
 import com.example.werewolfofthemillershollow.roles.Captain
 import com.example.werewolfofthemillershollow.roles.Role
+import com.example.werewolfofthemillershollow.roles.Villager
 import com.example.werewolfofthemillershollow.settings.App
 import com.example.werewolfofthemillershollow.settings.Icons
 import com.example.werewolfofthemillershollow.utility.*
@@ -59,7 +60,7 @@ class CaptainTurn(role : Role, var activity: GameActivity) : Turn<Role>(activity
 
     }
 
-    fun getWhoTalksInMorningOnClickHandler() : UsePowerDialog.OnClickListener{
+    fun getMorningOnClickHandler() : UsePowerDialog.OnClickListener{
 
         return object : UsePowerDialog.OnClickListener{
             override fun done(
@@ -306,6 +307,78 @@ class CaptainTurn(role : Role, var activity: GameActivity) : Turn<Role>(activity
                     adapter.emptyTargets()
                 }
             }
+    }
+
+    override fun getOnActionOnClickHandler(onAction: OnAction): UsePowerDialog.OnClickListener {
+        return object : UsePowerDialog.OnClickListener{
+            override fun done(
+                ability: Ability,
+                aliveList: ArrayList<Role>,
+                deadList: ArrayList<Role>,
+                adapter: TargetAdapter,
+                activity: GameActivity,
+                dialog: UsePowerDialog?
+            ): Boolean {
+                if (adapter.getTargets().isEmpty()){
+                    val alert = AlertDialog(text = R.string.should_use_power)
+                    alert.show(activity.supportFragmentManager, App.TAG_ALERT)
+                    return false
+                }
+
+                var touched : Role = Villager(activity)
+
+                for(index : Int in adapter.getTargets()){
+
+                    val target : Role = adapter.getList()[index]
+
+                    val i = activity.playerList.indexOf(target)
+
+                    if (i == -1)
+                        return false
+
+                    ability.use(self = getRole(), role = activity.playerList[i], activity.playerList)
+                    touched = activity.playerList[i]
+                    break
+
+                }
+
+                val onClick = object : AlertDialog.OnClick{
+                    override fun onClick(alertDialog: AlertDialog) {
+                        onAction.index ++
+                        onAction.onStart()
+                        dialog!!.dismiss()
+                        alertDialog.dismiss()
+                    }
+                }
+
+                val onTouch = object : AlertDialog.OnClick{
+                    override fun onClick(alertDialog: AlertDialog) {
+                        AlertDialog.displayDialog(
+                            activity = activity,
+                            text = R.string.good_night,
+                            rightButton = onClick,
+                            cancelable = false)
+                        alertDialog.dismiss()
+                    }
+                }
+
+                AlertDialog.displayDialog(
+                    activity = activity,
+                    text = -1,
+                    contentText = "${activity.getString(R.string.captain_touch)} (${activity.getString(R.string.touch)} ${touched.player})",
+                    rightButton = onTouch,
+                    cancelable = false)
+                return false
+            }
+
+            override fun reset(
+                aliveList: ArrayList<Role>,
+                deadList: ArrayList<Role>,
+                adapter: TargetAdapter
+            ) {
+                adapter.emptyTargets()
+            }
+        }
     }
 
     override fun shouldUsePower(gameActivity: GameActivity): Boolean {
