+++ app/src/main/java/com/example/werewolfofthemillershollow/utility/OnAction.kt
@@ -54,69 +54,13 @@ class OnAction(
             return
         }
 
-        val onClick = object : UsePowerDialog.OnClickListener{
-            override fun done(
-                ability: Ability,
-                aliveList: ArrayList<Role>,
-                deadList: ArrayList<Role>,
-                adapter: TargetAdapter,
-                activity: GameActivity,
-                dialog: UsePowerDialog?
-            ): Boolean {
-                if (adapter.getTargets().isEmpty()){
-                    val alert = AlertDialog(text = R.string.should_use_power)
-                    alert.show(activity.supportFragmentManager, App.TAG_ALERT)
-                    return false
-                }
-
-                for(index : Int in adapter.getTargets()){
-
-                    val target : Role = adapter.getList()[index]
-
-                    val i = activity.playerList.indexOf(target)
-
-                    if (i == -1)
-                        return false
-
-                    ability.use(self = current.getRole(), role = activity.playerList[i], activity.playerList)
-
-                }
-
-                val onClick = object : AlertDialog.OnClick{
-                    override fun onClick(alertDialog: AlertDialog) {
-                        index ++
-                        onStart()
-                        alertDialog.dismiss()
-                        dialog!!.dismiss()
-                    }
-                }
-
-                val goodNightDialog = AlertDialog(
-                    text = R.string.good_night,
-                    rightButton = onClick,
-                    cancelable = false)
-                goodNightDialog.show(activity.supportFragmentManager, App.TAG_ALERT)
-                return false
-            }
-
-            override fun reset(
-                aliveList: ArrayList<Role>,
-                deadList: ArrayList<Role>,
-                adapter: TargetAdapter
-            ) {
-                adapter.emptyTargets()
-            }
-        }
-
-
-
         val proceed = object : AlertDialog.OnClick{
             override fun onClick(alertDialog: AlertDialog) {
                 alertDialog.dismiss()
                 val dialog = UsePowerDialog(
                     turn = current,
                     ability = current.getOnCallAbility()!!,
-                    onClick = onClick,
+                    onClick = current.getOnActionOnClickHandler(this@OnAction),
                     onTargetClick = current.getOnTargetHandler(),
                     onDismissed = null,
                     cancelable = false,
